Mv test 15 into test 13 (test 15 was superflous)
